requirements.txt
django
djangorestframework
pandas
psycopg2-binary
celery
redis
openpyxl
